Fix conn string parsing with unsupported options
The ODBC specification does not mandate how the driver should deal with
unsupported options specified by client in the connection string.

When encountering an unsupported option, DuckDB driver registers a
diagnostic message in a format: 'Invalid keyword: <option_name>', that
can be fetched by client using SQLGetDiagRec function, and returns
SQL_SUCCESS_WITH_INFO instead of SQL_SUCCESS.

This is a reasonable approach, but there is inconsistent error handling
in its implementation in Connect::ParseInputStr. It only checks for
SQL_SUCCESS and treats SQL_SUCCESS_WITH_INFO as an error. This causes
the early exit from this function that prevents subsequent options in
the connection string to be processed. It also prevents the DSN name
to be set on the connection (even when the DSN option comes before the
unsupported one).

To fix this it is proposed to check for both SQL_SUCCESS and
SQL_SUCCESS_WITH_INFO and only do the early exit if some other error
code is returned by FindKeyValPair routine.

Testing: tests are included for a common option and for the DSN option.

Fixes: #59